# Return old recursion limit capabilities