--- a/test/Esprima.Tests/JavaScriptParserTests.cs
+++ b/test/Esprima.Tests/JavaScriptParserTests.cs
@@ -14,7 +14,7 @@ public class JavaScriptParserTests
 #if DEBUG
         const int Depth = 205;
 #else
-        const int Depth = 385;
+        const int Depth = 520;
 #endif
         var input = $"if ({new string('(', Depth)}true{new string(')', Depth)}) {{ }}";
         parser.ParseScript(input);