@adams85 2023-11-30
Wow, kudos! 🥇 

If I may ask, what tool did you use to analyze the code and track down that these methods are the culprits? Or did you just know by looking at the code? :)