@lahma 2023-11-30
I used debugger breakpoint hitting until stack was deep enough and started to read frames which could benefit from the attribute.