@adams85 2023-11-30
So, no black magic, just looking at the code mostly. I'm even more impressed... 😮